apply warnings cleanup

+++ src/coap_options.h
@@ -27,7 +27,7 @@
 
 //on addition of an option also change coap_options.C (array initializer) and count
 //Proxy-Uri
-extern size_t KNOWN_OPTIONS_COUNT;
+extern uint32_t KNOWN_OPTIONS_COUNT;
 extern uint16_t KNOWN_OPTIONS[];
 
 #define OPT_FLAG_CRITICAL 	(1u<<0u)
